Split each language to seperate file
